TINY-9601: Create DragImage module and improve drag image tests
