+++ modules/dragster/src/test/ts/atomic/datatransfer/DataTransferCloneTest.ts
@@ -6,6 +6,12 @@ import { assert } from 'chai';
 import { cloneDataTransfer, createDataTransfer, getDragImage } from 'ephox/dragster/datatransfer/DataTransfer';
 
 describe('atomic.dragster.datatransfer.DataTransferCloneTest', () => {
+  const dragImage = {
+    image: document.createElement('div'),
+    x: 10,
+    y: 20
+  };
+
   const getTestDataTransfer = (): DataTransfer => {
     const dataTransfer = createDataTransfer();
 
@@ -14,7 +20,7 @@ describe('atomic.dragster.datatransfer.DataTransferCloneTest', () => {
     dataTransfer.effectAllowed = 'copy';
     dataTransfer.setData('text/plain', '123');
     dataTransfer.setData('text/html', '<p>123</p>');
-    dataTransfer.setDragImage(document.createElement('div'), 10, 20);
+    dataTransfer.setDragImage(dragImage.image, dragImage.x, dragImage.y);
     dataTransfer.items.add(new window.File([ 'Lorem ipsum' ], 'test.txt', { type: 'text/plain' }));
 
     return dataTransfer;
@@ -36,7 +42,7 @@ describe('atomic.dragster.datatransfer.DataTransferCloneTest', () => {
     // setReadOnlyMode(clone);
 
     // Not have same items reference as original
-    assert.notStrictEqual(clone.items, original.items);
+    assert.notStrictEqual(clone.items, original.items, 'Clone should not have same items reference as original');
 
     // Same data as original
     assert.strictEqual(clone.dropEffect, original.dropEffect, 'Clone should have same dropEffect');
@@ -44,7 +50,14 @@ describe('atomic.dragster.datatransfer.DataTransferCloneTest', () => {
     assert.strictEqual(clone.getData('text/plain'), original.getData('text/plain'), 'Clone should have same data');
     assert.strictEqual(clone.getData('text/html'), original.getData('text/html'), 'Clone should have same data');
     assert.deepEqual(clone.types, original.types, 'Clone should have same types');
+
+    // Same drag image as original
+    const originalDragImage = getDragImage(original);
+    KAssert.eqSome('Original should have expected drag image', dragImage, originalDragImage);
+    const cloneDragImage = getDragImage(clone);
     KAssert.eqOptional('Clone should have same drag image data', getDragImage(original), getDragImage(clone));
+    originalDragImage.each((originalImage) =>
+      cloneDragImage.each((cloneImage) => assert.notStrictEqual(cloneImage, originalImage, 'Clone drag image should not have same reference as original')));
 
     // Same files as original
     const length = clone.files.length;
